update to 1.21.3
update loom to 1.8 and therefore gradle to 8.10
switched to name based HELD_TOOLS because the old method doesn't work anymore. This may cause issues if a player rejoins wit a name another player had previously, but generally this should be fine.

The generics look odd to me, but I don't know any better way how to extend the PlayerHeldItemFeatureRenderer than this.